fix(webpack): add rsbuild config to build dependencies
