+++ packages/core/src/cli/commands.ts
@@ -1,11 +1,8 @@
 import { join } from 'node:path';
 import { existsSync } from 'node:fs';
-import { color, isDev, logger } from '@rsbuild/shared';
+import { color, logger, type RsbuildMode } from '@rsbuild/shared';
 import { program, type Command } from '@rsbuild/shared/commander';
-import { loadEnv } from '../loadEnv';
-import { loadConfig, watchFiles } from '../config';
-import type { RsbuildMode } from '..';
-import { onBeforeRestartServer } from '../server/restart';
+import { init } from './init';
 
 export type CommonOptions = {
   config?: string;
@@ -28,82 +25,6 @@ export type InspectOptions = CommonOptions & {
 export type DevOptions = CommonOptions;
 
 export type PreviewOptions = CommonOptions;
-
-let commonOpts: CommonOptions = {};
-
-export async function init({
-  cliOptions,
-  isRestart,
-}: {
-  cliOptions?: CommonOptions;
-  isRestart?: boolean;
-}) {
-  if (cliOptions) {
-    commonOpts = cliOptions;
-  }
-
-  try {
-    const root = process.cwd();
-    const envs = loadEnv({
-      cwd: root,
-      mode: cliOptions?.envMode,
-    });
-
-    if (isDev()) {
-      onBeforeRestartServer(envs.cleanup);
-    }
-
-    const { content: config, filePath: configFilePath } = await loadConfig({
-      cwd: root,
-      path: commonOpts.config,
-      envMode: commonOpts.envMode,
-    });
-
-    const command = process.argv[2];
-    if (command === 'dev') {
-      const files = [...envs.filePaths];
-      if (configFilePath) {
-        files.push(configFilePath);
-      }
-
-      watchFiles(files);
-    }
-
-    const { createRsbuild } = await import('../createRsbuild');
-
-    config.source ||= {};
-    config.source.define = {
-      ...envs.publicVars,
-      ...config.source.define,
-    };
-
-    if (commonOpts.open && !config.dev?.startUrl) {
-      config.dev ||= {};
-      config.dev.startUrl = commonOpts.open;
-    }
-
-    if (commonOpts.host) {
-      config.server ||= {};
-      config.server.host = commonOpts.host;
-    }
-
-    if (commonOpts.port) {
-      config.server ||= {};
-      config.server.port = commonOpts.port;
-    }
-
-    return await createRsbuild({
-      cwd: root,
-      rsbuildConfig: config,
-    });
-  } catch (err) {
-    if (isRestart) {
-      logger.error(err);
-    } else {
-      throw err;
-    }
-  }
-}
 
 const applyCommonOptions = (command: Command) => {
   command
